ci: migrate to Cachix binary cache
This commit retires the self-hosted attic binary cache
used to speed up the build and deploy CI workflow as I
currently do not have the capacity to manage the binary
cache and associated authentication/authorization flows
dilligently. As such, I am migrating to a managed Cachix
cache for the build and deploy CI workflow.